chore(connect-web): refactor popupmanager

+++ packages/connect-webextension/webpack/inline-content-script.js
@@ -7,6 +7,9 @@ const buildResultPath = path.join(__dirname, '../build/trezor-connect-webextensi
 const buildResult = fs.readFileSync(buildResultPath, 'utf8');
 
 const replaceToken = '// <!--content-script-->';
+if (buildResult.indexOf(replaceToken) === -1) {
+    throw new Error('Content script token not found');
+}
 const updated = buildResult.replace(replaceToken, contentScript);
 
 fs.writeFileSync(buildResultPath, updated, 'utf8');
